mejorando diseño de fondo, reorganizando proyecto por carpetas, y comenzando componente texto
