+++ AzzAudio/src/components/main/Main.tsx
@@ -1,0 +1,11 @@
+import { Text } from "./layaout-main/sectionText/Text"
+
+export const Main = () => {
+    return (
+        <main className="h-52  flex-center 2xl:h-96 mb-96 w-full flex-center p-12 ">
+            <div className="w-full h-52 2xl:h-96  flex justify-center">
+                <Text/>
+            </div>
+        </main>
+    )
+}
