+++ AzzAudio/src/index.css
@@ -36,6 +36,7 @@
 
 
 @theme{
+    --color-primary:#020202;
     --color-secundary:  #00FFCC;
     --color-tertiary: #D9D9D9;
 }
